first seemingly working run with GP model
The interence result uses a GP with huge lengthscale such that GP latent
variables all have the same (wide) distribution. So the GP explains nothing and
all variance is explained by variability in athlete abilities. This is
surprising, because the prior over rho was relatively narrow with 90th
percentile at 24.31.