inference works on new, but not old records
Inference results appeared reasonable on the subset of automatically timed
records after the last changes, but the results didn't show the progression of
world record times as I intended, because the old records were missing. So I
included the old records in the data and reran inference. It turned out that it
didn't work with the expanded data set anymore with some chains not moving from
their initial values at all. So I tweaked the initialisation by shrinking the
ablity_std and increasing initial inconsistencies. This ensures that there is
sufficient probability mass around the record times so that the sampler has
reasonable gradient information.

With the new initialisation sampling took a long time (70 min), so appeared to
go well, but the resulting chains could be quite different, i.e., sampling was
highly dependent on initialisation. Each individual chain appeared to have
converged to a reasonable result, but the sampler did apparently not properly
explore the regions of the posterior. One chain particularly stuck out: in it
the Gaussian process was unused with alpha ~ 0. At the same time it reached on
average the highest lp-values.

I probably need to find priors that allow better mixing by better connecting
the different regions of the posterior.