fully working inference after several tweaks
changes:
1. simpler, Stan built-in likelihood as previously tested
2. increasing / widen prior of rho
3. explicitly centre the average around 10
4. use lognormal prior for astds
5. widening priors for abilities and ability_std

I believe that 1. and 3. had no influence on the success of the inference, but
changing the priors had. The change for rho (2.) shouldn't have been big, but
biases rho more towards larger values and covers a wider range of values. My
intuition is that swapping the half-normal prior on astds against a lognormal
was the crucial step. Perhaps not the particular shape of the prior had the
largest impact, but that it became quite a bit wider. Widening the priors for
abilities and ability_std could also have had a positive impact on inference.
After all that was the plan after seeing that the sampler got stuck in different
regions of the posterior.